feat: add SessionManager
* Adds loading, error, and time formatting utils

Implements reusable components for displaying loading indicators and error messages.

Introduces a utility function for formatting dates into a user-friendly time format.

* Adds admin game management components

Introduces components for managing games in the admin panel.

This includes:
- A card component displaying game details with edit and delete options.
- A list component to display multiple game cards in a grid layout.
- A slot management component for adding and removing time slots for games.

* Adds game management page

Implements a new page for managing escape room games.

This includes the ability to create, edit, and delete games, along with their associated slots (time availability). The page provides a form for game details and a list of existing games for administration.

---------

Co-authored-by: Enzo Borges Pereira <enzo077@outlook.fr>